Restructure matching code
